BackEnd/requirements.txt
fastapi[standard]
uvicorn[standard]
python-dotenv
supabase
python-jose